+++ addons/wms_location/tests/__init__.py
@@ -2,6 +2,7 @@ from . import (
     test_barcode_format,
     test_barcode_integrity,
     test_capacity,
+    test_fpat_fx1,
     test_location,
     test_location_delete,
     test_polyomino,